requirements.txt
aiohappyeyeballs==2.3.5
aiosignal==1.3.1
aiosqlite==0.20.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.112.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.7
multidict==6.0.5
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
yarl==1.18.3
sqlalchemy==2.0.38
alembic==1.15.1
aiohttp==3.11.13
pandas==2.2.3
requests==2.32.3
openpyxl==3.1.5
paramiko==3.5.0
email-validator==2.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6